refactor GetRecipesResponseModel and seperate ingredient model to new class
