+++ lib/features/category_recipes/presentation/cubit/get_recipes_by_category/get_recipes_by_category_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:looqma/core/common/models/get_recipes_query_model.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/home/data/repos/home_repo.dart';
 
 part 'get_recipes_by_category_cubit.freezed.dart';
