+++ lib/features/home/data/repos/home_repo.dart
@@ -5,7 +5,7 @@ import 'package:looqma/core/networking/api_service.dart';
 import 'package:looqma/features/home/data/models/all_categories_model.dart';
 import 'package:looqma/features/home/data/models/all_countries_model.dart';
 import 'package:looqma/core/common/models/get_recipes_query_model.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 
 class HomeRepo {
   final ApiService _apiService;
