+++ lib/features/home/presentation/cubit/get_recipes/get_new_recipes/get_new_recipes_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:looqma/core/common/models/get_recipes_query_model.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/home/data/repos/home_repo.dart';
 
 part 'get_new_recipes_cubit.freezed.dart';
