+++ lib/features/home/presentation/views/widgets/show_recipes_by_country_loading.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_constants.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/home/presentation/views/widgets/recipe_item.dart';
 import 'package:skeletonizer/skeletonizer.dart';
