+++ lib/features/recipe_details/presentation/views/recipe_details_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_colors.dart';
 import 'package:looqma/core/utils/app_styles.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/dropdown_button.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/recipe_image.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/view_recipe_ingrediants_and_procedure.dart';
