+++ lib/features/recipe_details/presentation/views/widgets/ingredients_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/ingredients_item.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/serve_numbers.dart';
