+++ lib/features/recipe_details/presentation/views/widgets/view_recipe_ingrediants_and_procedure.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_colors.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/ingredients_list_view.dart';
 import 'package:looqma/features/recipe_details/presentation/views/widgets/procedure_list_view.dart';
