+++ lib/features/saved_recipe/data/repos/saved_recipes_repo.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:looqma/core/networking/api_error.dart';
 import 'package:looqma/core/networking/api_result.dart';
 import 'package:looqma/core/networking/api_service.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 
 class SavedRecipesRepo {
   final ApiService _apiService;
