+++ lib/features/saved_recipe/presentation/views/widgets/saved_recipe_details.dart
@@ -5,7 +5,7 @@ import 'package:looqma/core/common/recipe_save_toggle/cubit/recipe_save_toggle_c
 import 'package:looqma/core/common/widgets/save_recipe_button.dart';
 import 'package:looqma/core/utils/app_colors.dart';
 import 'package:looqma/core/utils/app_styles.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/saved_recipe/presentation/cubit/get_saved_recipes/get_saved_recipes_cubit.dart';
 
 class SavedRecipeDetails extends StatelessWidget {
