+++ lib/features/saved_recipe/presentation/views/widgets/saved_recipe_item.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/common/widgets/custom_rate.dart';
 import 'package:looqma/core/utils/app_colors.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/saved_recipe/presentation/views/widgets/saved_recipe_details.dart';
 
 class SavedRecipeItem extends StatelessWidget {
