+++ lib/features/saved_recipe/presentation/views/widgets/saved_recipe_loading_listview.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_constants.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/saved_recipe/presentation/views/widgets/saved_recipe_item.dart';
 import 'package:skeletonizer/skeletonizer.dart';
