+++ lib/features/search_recipes/data/repos/search_repo.dart
@@ -3,7 +3,7 @@ import 'package:looqma/core/common/models/get_recipes_query_model.dart';
 import 'package:looqma/core/networking/api_error.dart';
 import 'package:looqma/core/networking/api_result.dart';
 import 'package:looqma/core/networking/api_service.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 
 class SearchRepo {
   SearchRepo(this._apiService);
