+++ lib/features/search_recipes/presentation/cubit/search_recipe/search_recipe_cubit.dart
@@ -6,7 +6,7 @@ import 'package:flutter_debouncer/flutter_debouncer.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:looqma/core/common/models/get_recipes_query_model.dart';
 import 'package:looqma/core/common/widgets/show_toast.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/search_recipes/data/repos/search_repo.dart';
 
 part 'search_recipe_cubit.freezed.dart';
