+++ lib/features/search_recipes/presentation/views/widgets/search_result.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:looqma/features/home/data/models/get_recipes_response_model.dart';
+import 'package:looqma/core/common/models/get_recipes_response_model.dart';
 import 'package:looqma/features/search_recipes/presentation/views/widgets/search_grid_view.dart';
 import 'package:looqma/features/search_recipes/presentation/views/widgets/search_result_length.dart';
